requirements.txt
# FastAPI Framework
fastapi==0.115.6
uvicorn[standard]==0.34.0

# Database
sqlalchemy==2.0.36
pymysql==1.1.0
cryptography==42.0.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Validation & Settings
pydantic==2.10.5
pydantic-settings==2.7.1
python-dotenv==1.0.0

# HTTP Client (for WeChat API)
httpx==0.26.0

# File Upload
python-multipart==0.0.6
aiofiles==23.2.1

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
hypothesis==6.92.2
httpx==0.26.0

# Utilities
python-dateutil==2.8.2
